backend/requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
google-genai
python-dotenv
pydantic
requests
